requirements.txt
# Text2SQL 智能体系统依赖

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# 向量检索
faiss-cpu>=1.7.0
numpy>=1.21.0

# 数据库
pymysql>=1.1.0

# HTTP 请求 (Ollama API)
requests>=2.28.0

# 可选: 如果使用其他 Embedding
# sentence-transformers>=2.2.0

# 环境变量加载
python-dotenv>=1.0.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
